fix: category comments' location
